feat: import tokens redesign (#21704)
## **Description**

This ticket aims to redesign the import token flow.
Here is the ticket:
https://consensyssoftware.atlassian.net/browse/MMASSETS-62
And Figma:
https://www.figma.com/file/TfVzSMJA8KwpWX8TTWQ2iO/MMASSETS-15%3A-Asset-list-and-details?type=design&node-id=1491-56878&mode=design

Notes: 
1- In the Figma designs you will notice that the notification you get
when you import a token is a black box notif; but the one in this PR is
a green box; we have decided to eventually go with a notification
component that already existed.
2- In the Figma design (2.0A) the confirmation page should also show the
balance in Fiat. We decided to split that part in a separate ticket and
for this PR use the same tokenBalance component used in the old
confirmation modal.

## **Related issues**

Fixes: #https://github.com/MetaMask/metamask-extension/issues/21052

## **Manual testing steps**

1. Go to home page
2. Click on import tokens button
3. try searching for your token
4. Select the tokens you need
5. Click Next
6. If you want to exit at this point, click the "x" and confirm; and if
you want to go back click "back"
7. If you clicked back on step 6; click import and you should be
redirected to home page.
8. You should be able to see your tokens on home page.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**


https://github.com/MetaMask/metamask-extension/assets/10994169/858b92aa-0ad4-48d1-b71f-b8ec90f62124

### **After**


https://github.com/MetaMask/metamask-extension/assets/10994169/0a5f8d4d-39c0-45fb-a76e-35f278e9070a


## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.