[e2e] Multiple provider e2e test implementation. (#21571)
## **Description**

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

This is an basic implementation of eip6963 e2e tests checks that copies
the `dist/` folder when the script is ran and a conditional for a
the`process.env.MULTIPROVIDER` as to not mess with the other tests that
require only one provider. I went with this approach as it was the
lowest effort implementation. If this is not the direction we want to
go, whether we handle it via `--build-type` or all within the ci, open
to those as well. <strike>This will require a bump of the `test-dapp`
version. (done)</strike>

## **Related issues**

Fixes: https://github.com/MetaMask/MetaMask-planning/issues/1500

## **Manual testing steps**

1. Run `yarn && yarn build:test && yarn test:e2e:chrome:multi-provider`
(node v18). Might have to update chromedriver to either v117/v118.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [x] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Howard Braham <howrad@gmail.com>
Co-authored-by: Alex Donesky <adonesky@gmail.com>