fix(CI): Fix shallow checkout (#22558)
## **Description**

The shallow checkout step in CI has been checking out the wrong commit
in many workflows. It worked correctly for builds from forks, but builds
from this repository would use the latest commit from the given branch,
rather than the specific commit that the workflow was triggered for. And
for tags, it was using the correct commit, but the checkout would fail
if the commit wasn't what the tag currently pointed at.

The shallow checkout script has been updated to use a simpler set of
commands which are the same for builds on this repository and on forks.
The new checkout script will always checkout the current commit,
represented by the `CIRCLE_SHA1` environment variable.

## **Related issues**

Fixes #22579

## **Manual testing steps**

See that the checkout step works correctly in the CI workflow for this
PR, and that it has the expected console message.

The behavior for a fork can be seen here:
https://github.com/MetaMask/metamask-extension/pull/22561

The behavior for a tag is not easily tested because we don't trigger
CircleCI workflows from tags, but you can run the git commands locally
to see that they work.

## **Screenshots/Recordings**

N/A

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've clearly explained what problem this PR is solving and how it
is solved.
- [x] I've linked related issues
- [x] I've included manual testing steps
- [x] I've included screenshots/recordings if applicable
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.
- [x] I’ve properly set the pull request status:
  - [x] In case it's not yet "ready for review", I've set it to "draft".
- [x] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.