Set FONTCONFIG_PATH on CircleCI containers (#22657)
## **Description**

Some test jobs in CI occasionally fail with the following error message:

    Fontconfig error: cannot load default config file

Why this happens only occasionally is a mystery, especially considering
that we are running Chromium via Chromedriver in sandbox mode. But the
most common solution for this kind of problem (for example,
[here](https://askubuntu.com/questions/1151094/cannot-load-config-file-from-etc-fonts-fonts-conf))
is to set the `FONTCONFIG_PATH` environment variable to point to the
directory that contains the `fontconfig` configuration file. Given that
context, this commit sets `FONTCONFIG_PATH` to `/etc/fonts`. I've
shelled into the Docker image that we use in CI to confirm that 1) this
environment variable is not set and 2) `/etc/fonts` exists.

## **Related issues**

Fixes: #22564

## **Manual testing steps**

There is no way to test this, but if this works, no jobs in CI should
time out and print the error message above.